requirements.txt
line-bot-sdk
bs4
flask
pymongo
matplotlib
datetime
imgurpython
pandas
pandas_datareader
fix_yahoo_finance == 0.0.22
mpl_finance
html5lib